Move create_volume_history and use it for tests Add volume validator constant
